Use position sticky for the headers (#719)
* Rename TableContent to TableRevisionContent for more clarity

* Split the ResultsTable into 2 different components

* Extract the Results Tables controls to its own component

* Move the Suspended code down to ResultsTable

* Move the search, framework and revision dropdowns, and the download button, closer to the table

* Adjust the styling following the move of the table top bar

* Use position: sticky on the table headers
